requirements.txt
mcp>=0.9.0
google-genai>=1.0.0
python-dotenv>=1.0.0
litellm>=1.50.0
aiohttp>=3.9.0
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
httpx>=0.25.0
huggingface_hub
Pillow
replicate>=1.0.0
groq>=0.4.0
fal-client>=0.5.0
duckduckgo-search>=6.0.0
elevenlabs>=1.0.0
gunicorn==21.2.0

